Final commit of what was sent to Jeff Blackmon

+++ CAENPhaChannelParameters.h
@@ -51,7 +51,7 @@ public:
 
   // <EnergyFilter> parameters:
 
-  unsigned decayTime;
+  double    decayTime;
   double   trapRiseTime;
   double   flattopDelay;
   double   trapFlatTop ;
